FIX: Wrong RGB status values
Changed variables to uint16 instead of uint8
Brightness value needs to be scaled to initialize HSV value

+++ inc/Channels/RGBChannel.h
@@ -23,8 +23,8 @@ class RGBChannel : public LightChannel
   
     const std::string name() override;
 
-    uint8_t _lastHueValue = 0;
-    uint8_t _lastSatValue = 0;
+    uint16_t _lastHueValue = 0;
+    uint16_t _lastSatValue = 0;
 
     DimmableValue<uint16_t> _hue;
     DimmableValue<uint16_t> _saturation;
